Rename type inference agent file.
