+++ rift-engine/rift/agents/__init__.py
@@ -1,10 +1,9 @@
 from .abstract import *
-from .add_missing_types import *
+from .type_inference_agent import *
 from .aider_agent import *
 from .code_edit import *
 from .curl_agent import *
 from .engineer import *
-from .mentat_agent import *
 from .registry import *
 from .rift_chat import *
 from .smol import *
